后端改了很多

+++ backend/springboot/src/main/java/com/example/demo/controller/FavController.java
@@ -5,13 +5,12 @@ import cn.hutool.json.JSONObject;
 //import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.example.demo.common.Result;
 //import com.example.demo.entity.Fav;
+import com.example.demo.mapper.FavMapper;
+import com.example.demo.mapper.UserMapper;
 import com.example.demo.service.impl.FavServiceImpl;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
 //import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Resource;
 //import java.util.ArrayList;
@@ -32,10 +31,20 @@ public class FavController {
     FavServiceImpl favService;
 
     @Resource
+    FavMapper favMapper;
+
+    @Resource
     ObjectController objectController;
 
     @PostMapping("collect")
-    public Result<?> collect(Integer favId, Integer objectId) {
+    public Result<?> collect(@RequestParam("favId") Integer favId, @RequestParam("objectId") Integer objectId) {
+        if (favId != 1 || objectId < 0) {
+            return Result.error("-1", "收藏失败");
+        }
+
+        if (favService.isExist(favId, objectId) != 0) {
+            return Result.error("-1", "已收藏，收藏失败");
+        }
         Integer count = favService.collect(favId, objectId);
         if (count <= 0) {
             return Result.error("-1", "收藏失败");
@@ -45,7 +54,14 @@ public class FavController {
     }
 
     @GetMapping("deleteCollection")
-    public Result<?> deleteCollection(Integer favId, Integer objectId) {
+    public Result<?> deleteCollection(@RequestParam(value = "favId", required = true) Integer favId, @RequestParam(value = "objectId", required = true) Integer objectId) {
+        if (favId < 0 || objectId < 0) {
+            return Result.error("-1", "数据有误，删除失败");
+        }
+
+        if (favService.isExist(favId, objectId) == 0) {
+            return Result.error("-1", "未收藏，删除失败");
+        }
         Integer count = favService.deleteCollection(favId, objectId);
         if (count <= 0) {
             return Result.error("-1", "收藏删除失败");
